chore(suite): split rows and targets into separate file
(cherry picked from commit 04112a59897522e52be1b9eb5f3c0f99da4bd425)
